create User model from Article
